update demo files
